added middleware auth

+++ backend/controllers/sauce.js
@@ -76,17 +76,31 @@ exports.modifySauce = (req, res, next) => {
   );
 };
 exports.deleteSauce = (req, res, next) => {
-  Sauce.deleteOne({_id: req.params.id}).then(
-    () => {
-      res.status(200).json({
-        message: 'Deleted!'
-      });
-    }
-  ).catch(
-    (error) => {
-      res.status(400).json({
-        error: 'Failed to delete'
-      });
+  Sauce.findOne({_id: req.params.id}).then(
+    (sauce) => {
+      if (!sauce) {
+        return res.status(404).json({
+          error: new Error('no such thing')
+        });
+      }
+      if(sauce.userId !== req.auth.userId) {
+        return res.status(401).json({
+          error: new Error('Unauthorized request.')
+        });
+      }
+      Sauce.deleteOne({_id: req.params.id}).then(
+        () => {
+          res.status(200).json({
+            message: 'Deleted!'
+          });
+        }
+      ).catch(
+        (error) => {
+          res.status(400).json({
+            error: 'Failed to delete'
+          });
+        }
+      );
     }
   );
 };
